app/requirements.txt
python-dotenv==0.21.1
uvicorn==0.20.0
fastapi==0.92.0
openai==0.27.0
langchain==0.0.98
unstructured==0.5.1
chromadb==0.3.10